+++ src/index.js
@@ -14,6 +14,7 @@ const Client = require('./client');
 const RequestExecutor = require('./request-executor');
 const { DefaultRequestExecutor } = require('./default-request-executor');
 const { Collection } = require('./collection');
+const MemoryStore = require('./memory-store');
 
 module.exports = Object.assign(
   {}, {
@@ -21,6 +22,7 @@ module.exports = Object.assign(
     RequestExecutor,
     DefaultRequestExecutor,
     Collection,
+    MemoryStore,
   },
   require('./generated'),
 );