+++ test/it/user-linked-object.ts
@@ -4,7 +4,8 @@ import {
   Collection,
   DefaultRequestExecutor,
   LinkedObject,
-  User
+  User,
+  ResponseLinks
 } from '@okta/okta-sdk-nodejs';
 import utils = require('../utils');
 import getMockLinkedObject = require('./mocks/linked-object');
@@ -60,7 +61,7 @@ describe('User linked object API', () => {
   });
 
   describe('Get linked object value', () => {
-    let links;
+    let links: Collection<ResponseLinks>;
     beforeEach(async () => {
       await client.userApi.setLinkedObjectForUser({
         associatedUserId: associateUser.id,
@@ -76,9 +77,9 @@ describe('User linked object API', () => {
       });
       expect(links).to.be.instanceOf(Collection);
       await links.each(link => {
-        // OKTA-512349: ResponseLinks is not specified in getLinkedObjectsForUser signature
-        // expect(link).to.be.instanceOf(ResponseLinks);
-        expect(link._links.self.href).contains(primaryUser.id);
+        expect(link).to.be.instanceOf(ResponseLinks);
+        // OKTA-512349: ResponseLinks has empty schema
+        expect(link['_links'].self.href).contains(primaryUser.id);
       });
     });
 
@@ -89,8 +90,8 @@ describe('User linked object API', () => {
       });
       expect(links).to.be.instanceOf(Collection);
       await links.each(link => {
-        // expect(link).to.be.instanceOf(ResponseLinks);
-        expect(link._links.self.href).contains(associateUser.id);
+        expect(link).to.be.instanceOf(ResponseLinks);
+        expect(link['_links'].self.href).contains(associateUser.id);
       });
     });
   });