+++ test/type/memory-store.test-d.ts
@@ -4,5 +4,5 @@ import { MemoryStore } from '../../src/types/memory-store';
 
 const memoryStore = new MemoryStore({});
 expectType<Promise<void>>(memoryStore.delete('key'));
-expectType<Promise<string>>(memoryStore.get('key'));
+expectType<Promise<string|undefined>>(memoryStore.get('key'));
 expectType<Promise<string>>(memoryStore.set('key', 'value'));